Updating the website
Adding community links (metafetish.com and metafetish.club)
Adding a placeholder software page ("We're working on the first
release")
Updated the dev docs link to point to the protocol docs (in lew of
anything better)
Added a hardware overview page (to contain the hardware support table).
Removed a bunch of unused files that were confusing me.